# diploma_work/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Structure API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/challenges/store": {
            "post": {
                "tags": [
                    "User Challenges"
                ],
                "summary": "Store a new user challenge",
                "description": "Store a new user challenge.",
                "operationId": "2b65eebf77f6ffdb0730cb0c770d7b5e",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "example": "Bearer your_token"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserChallengeResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserChallengeResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/user/challenges/show/{user}": {
            "get": {
                "tags": [
                    "User Challenges"
                ],
                "summary": "Show user challenges",
                "description": "Show the challenges for a specific user.",
                "operationId": "1554eec7ab6028cb50afa3aac81f89c3",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "example": "Bearer your_token"
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to show challenges for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "daily": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserChallengeResource"
                                            }
                                        },
                                        "weekly": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserChallengeResource"
                                            }
                                        },
                                        "monthly": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserChallengeResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/user/challenges/delete/{user}": {
            "delete": {
                "tags": [
                    "User Challenges"
                ],
                "summary": "Delete user challenges",
                "description": "Remove challenges for a specific user.",
                "operationId": "1d8282fba1e5c4ee9a3d25e974a78880",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "example": "Bearer your_token"
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user whose challenges will be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User challenges deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChallengeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "10k Run Challenge"
                    },
                    "description": {
                        "type": "string",
                        "example": "Complete a 10km run within a month."
                    },
                    "due_type": {
                        "type": "string",
                        "example": "monthly"
                    },
                    "challenge_type": {
                        "type": "string",
                        "example": "distanceChallenge"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "points": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
            "UserChallengeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "challenge": {
                        "$ref": "#/components/schemas/ChallengeResource"
                    },
                    "challenge_status": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}